I need to rework the whole DataSourceAttributesConfiguration and DataSourceAttributesInstanceConfiguration stuff. It's going to get a bit tricky so a checkpoint commit now, even though this doesn't even compile, is needed.

+++ src/com/obtuse/ui/layout/play/GoldenRatioLayoutManager.java
@@ -533,7 +533,7 @@ public class GoldenRatioLayoutManager implements LayoutManager {
 
     private static int round( final double d ) {
 
-        return (int)Math.floor( (double)( d + 0.5F ) );
+        return (int)Math.floor( d + 0.5F );
 
     }
